+++ zcash_client_backend/src/data_api.rs
@@ -353,6 +353,8 @@ pub struct WalletSummary<AccountId: Eq + Hash> {
     fully_scanned_height: BlockHeight,
     scan_progress: Option<Ratio<u64>>,
     next_sapling_subtree_index: u64,
+    #[cfg(feature = "orchard")]
+    next_orchard_subtree_index: u64,
 }
 
 impl<AccountId: Eq + Hash> WalletSummary<AccountId> {
@@ -362,14 +364,17 @@ impl<AccountId: Eq + Hash> WalletSummary<AccountId> {
         chain_tip_height: BlockHeight,
         fully_scanned_height: BlockHeight,
         scan_progress: Option<Ratio<u64>>,
-        next_sapling_subtree_idx: u64,
+        next_sapling_subtree_index: u64,
+        #[cfg(feature = "orchard")] next_orchard_subtree_index: u64,
     ) -> Self {
         Self {
             account_balances,
             chain_tip_height,
             fully_scanned_height,
             scan_progress,
-            next_sapling_subtree_index: next_sapling_subtree_idx,
+            next_sapling_subtree_index,
+            #[cfg(feature = "orchard")]
+            next_orchard_subtree_index,
         }
     }
 
@@ -403,6 +408,13 @@ impl<AccountId: Eq + Hash> WalletSummary<AccountId> {
     /// roots passed to [`WalletCommitmentTrees::put_sapling_subtree_roots`].
     pub fn next_sapling_subtree_index(&self) -> u64 {
         self.next_sapling_subtree_index
+    }
+
+    /// Returns the Orchard subtree index that should start the next range of subtree
+    /// roots passed to [`WalletCommitmentTrees::put_orchard_subtree_roots`].
+    #[cfg(feature = "orchard")]
+    pub fn next_orchard_subtree_index(&self) -> u64 {
+        self.next_orchard_subtree_index
     }
 
     /// Returns whether or not wallet scanning is complete.